+++ kedro_kubeflow/vertex_ai/client.py
@@ -119,14 +119,16 @@ class VertexAIPipelinesClient:
     def schedule(
         self,
         pipeline,
-        image,
+        experiment_name,
+        experiment_namespace,
         cron_expression,
         image_pull_policy="IfNotPresent",
     ):
         """
         Schedule pipeline to Vertex AI with given cron expression
         :param pipeline:
-        :param image:
+        :param experiment_name:
+        :param experiment_namespace:
         :param cron_expression:
         :param image_pull_policy:
         :return:
@@ -136,7 +138,7 @@ class VertexAIPipelinesClient:
         ) as spec_output:
             self.compile(
                 pipeline,
-                image,
+                self.run_config.image,
                 output=spec_output.name,
                 image_pull_policy=image_pull_policy,
             )